*: rename ec2 -> aws and gce -> gcp
Consistently use the name of the cloud platform rather than of its
compute service.

Temporarily accept "ec2" as a compatibility alias for "aws" to ease
migration.

Not for backport to spec2x.
